Refactored structure, fixed imports and routing
