+++ software_side/walkbuddy_reactNative/frontend_reactNative/app/(tabs)/places.tsx
@@ -18,7 +18,7 @@ import {
   toggleFavourite,
   markUsed,
   PlaceItem,
-} from "../lib/placesStore";
+} from "../../src/utils/placesStore"
 
 /*
   TEMPORARY TESTING ONLY
