+++ software_side/walkbuddy_reactNative/frontend_reactNative/app/_layout.tsx
@@ -5,8 +5,8 @@ import { StatusBar } from "expo-status-bar";
 import "react-native-reanimated";
 
 import { useColorScheme } from "@/hooks/use-color-scheme";
-import { CurrentLocationProvider } from "./lib/locationSaver";
-import { SessionProvider } from "./SessionContext";
+import { CurrentLocationProvider } from "../src/utils/locationSaver";
+import { SessionProvider } from "../src/context/SessionContext";
 
 export const unstable_settings = {
   anchor: "(tabs)",
@@ -25,10 +25,6 @@ export default function RootLayout() {
             }}
           >
             <Stack.Screen name="(tabs)" />
-            <Stack.Screen
-              name="modal"
-              options={{ presentation: "modal" }}
-            />
           </Stack>
           <StatusBar style="auto" />
         </ThemeProvider>
