+++ software_side/walkbuddy_reactNative/frontend_reactNative/src/utils/favourites.tsx
